refactor(harness): consolidate ruby-pay-kit-server into the canonical ruby adapter (Ludo r3311413501)
The dual-protocol PayKit adapter is a strict superset of the previous
MPP-only ruby-server, so there is no reason to carry two distinct
harness impls. Drop the old harness/ruby-server/server.rb (189 lines,
Mpp-only) and rename harness/pay-kit-server/server.rb to
harness/ruby-server/server.rb. The impl id collapses from
{ruby, ruby-pay-kit-server} to a single `ruby` that supports both
intents.

- harness/src/implementations.ts: drop the separate ruby-pay-kit-server
  entry, give the canonical `ruby` impl the dual-protocol label and
  intents: ["charge", "x402-exact"].
- harness/src/intents/charge.ts: every scenario serverIds list that
  previously named ruby-pay-kit-server now names ruby (push,
  network-mismatch, cross-route-replay, decimals-9, sol-native,
  cross-server-portability, idempotent-resubmit).
- harness/test/e2e.test.ts: comment update.
- .github/workflows/ruby.yml: job renamed to interop-ruby, drops
  PAY_KIT_INTEROP_SERVERS env (MPP_INTEROP_SERVERS=ruby,typescript
  enables the cross-server-portability pair), test name pattern
  switched to "ruby".
- harness/ruby-server/server.rb: ready payload `implementation` field
  is now "ruby" and the error-prefix strings stop saying
  "pay-kit-server".
- ruby/test/pay_kit/harness_adapter_test.rb: ADAPTER path and ready
  payload assertion follow the rename.

Local matrix: 17 scenarios still green (8 charge + 1 x402-exact + 2
cross-server + 1 idempotent + the 5 added in 340fb3b).
