+++ harness/src/implementations.ts
@@ -170,14 +170,21 @@ export const serverImplementations: ImplementationDefinition[] = [
   },
   {
     id: "ruby",
-    label: "Ruby HTTP server",
+    label: "Ruby PayKit server (dual protocol)",
     role: "server",
+    // One adapter binary, two settle paths. The harness orchestrator
+    // sets either `X402_INTEROP_*` (x402-exact intent) or `MPP_INTEROP_*`
+    // (charge intent); the adapter detects which one is active (via
+    // PAY_KIT_INTEROP_PROTOCOL hint or namespace probe) and routes
+    // through PayKit::Rack::Dispatcher (x402) or Mpp::Server::Charge
+    // directly (mpp).
     command: [
       "sh",
       "-c",
       "cd ../ruby && bundle exec ruby ../harness/ruby-server/server.rb",
     ],
     enabled: isEnabled("ruby", "MPP_INTEROP_SERVERS", false),
+    intents: ["charge", "x402-exact"],
   },
   {
     id: "lua",
@@ -258,25 +265,5 @@ export const serverImplementations: ImplementationDefinition[] = [
     ],
     enabled: isEnabled("ruby-x402-server", "X402_INTEROP_SERVERS", false),
     intents: ["x402-exact"],
-  },
-  {
-    id: "ruby-pay-kit-server",
-    label: "Ruby PayKit server (dual protocol)",
-    role: "server",
-    // One adapter binary, two settle paths. The harness orchestrator
-    // sets either `X402_INTEROP_*` (for the x402-exact intent) or
-    // `MPP_INTEROP_*` (for the charge intent); pay-kit-server detects
-    // which one is active and routes through PayKit::Rack::Dispatcher.
-    // This is the cross-language proof of the dual-protocol PayKit
-    // surface.
-    command: [
-      "sh",
-      "-c",
-      "cd ../ruby && bundle exec ruby ../harness/pay-kit-server/server.rb",
-    ],
-    // Defaults off; opt-in via `PAY_KIT_INTEROP_SERVERS=ruby-pay-kit-server`.
-    // The CI workflow flips this on for both protocols.
-    enabled: isEnabled("ruby-pay-kit-server", "PAY_KIT_INTEROP_SERVERS", false),
-    intents: ["charge", "x402-exact"],
   },
 ];
